requirements.txt
openvr==1.16.0801
numpy>=1.26.0
pygame==2.5.2
pyopengl==3.1.7
mojimoji==0.0.12 